:sparkles: feat: Tratamento de Exceções personalizado

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,14 +1,20 @@
 package com.devsuperior.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dtos.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -18,7 +24,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product obj = repository.findById(id).get();
+        Product obj = repository.findById(id).orElseThrow(() ->
+        new ResourceNotFoundException(id));
         return new ProductDTO(obj);
     }
 
@@ -35,10 +42,15 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id,ProductDTO obj){
-        Product entity = repository.getReferenceById(id);
-        updateData(entity,obj);
-        repository.save(entity);
-        return new ProductDTO(entity);
+        try{
+            Product entity = repository.getReferenceById(id);
+            updateData(entity,obj);
+            repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch(EntityNotFoundException e){
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(Product entity, ProductDTO obj) {
@@ -48,8 +60,16 @@ public class ProductService {
         entity.setImgUrl(obj.imgUrl());
     }
 
-    @Transactional
-    public void deleteById(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+	    if (!repository.existsById(id)) {
+		    throw new ResourceNotFoundException(id);
+	    }
+	    try {
+        	repository.deleteById(id);    		
+	    }
+    	catch (DataIntegrityViolationException e) {
+        	throw new DatabaseException("Falha de integridade referencial");
+   	    }
     }
 }
